ci-matrix: Make file pruning safer
Accept a glob pattern instead of a directory so that the tool only reads
files that match the pattern. Should generally only be used with a glob
pattern that matches file extensions or file names.

Rename the flag to timing-files

Remove the default max age, to prevent accidentaly deleting files.

Add test coverage for the function.